pyproject.toml
[project]
name = "test-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic>=2.12.2",
    "pydantic-ai>=1.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
    "pypdf2>=3.0.0",
    "python-multipart>=0.0.9",
    "logging>=0.4.9.6",
    "requests>=2.31.0",
]
